thumbnail/requirements.txt
scenedetect
moviepy
pillow
